wrappers: parse_*: file_name uses same bound as miette
The three parse functions, `parse_ast`, `parse`, and
`parse_with_context` take `file_name` as the first argument.
This argument is used to construct `miette::NamedSource`,
the `new` constructor of which takes `AsRef<str>`.
This change ensures that `miette`'s choice bound is preserved.

Unfortunately, `miette` misuses the bound, and clones (to `String`)
in library code which is unidiomatic. Perhaps `miette` does not need to
clone, and they knew this when writing the function signature? Does a
new version of `miette` address this?